refactor: 💡 ruler utils
