+++ packages/charts-utils/src/scale/format-scale-label/format-scale-label.test.ts
@@ -1,26 +1,26 @@
+import {
+  register as registerTimezone,
+  unregister as unregisterTimezone,
+} from 'timezone-mock';
+
 import formatScaleLabel from './format-scale-label';
 
-const RealDate = Date;
-
 beforeAll(() => {
-  global.Date = jest.fn().mockImplementation(date => new RealDate(date));
-  global.Date.UTC = jest
-    .fn()
-    .mockImplementation(date => new RealDate(date).getUTCDate());
+  registerTimezone('UTC');
 });
 
 afterAll(() => {
-  global.Date = RealDate;
+  unregisterTimezone();
 });
 
-test('returns original value', () => {
+test('returns original scale value', () => {
   const value = 'marketing';
   const result = formatScaleLabel(value, { type: 'band' });
 
   expect(result).toEqual(value);
 });
 
-test('applies scale settings formatter function', () => {
+test('applies formatter function from scale settings ', () => {
   const value = 'formatted';
   const formatFunction = () => 'formatted';
   const result = formatScaleLabel(100, {
@@ -32,8 +32,10 @@ test('applies scale settings formatter function', () => {
 });
 
 test('formats "Date" instance as ISO string', () => {
-  const value = new Date('2020-09-12:12:00:00Z');
+  const value = new Date('2020-09-12T12:00:00Z');
   const result = formatScaleLabel(value);
 
-  expect(result).toMatchInlineSnapshot(`2020-09-12T12:00:00.000Z`);
+  expect(result).toMatchInlineSnapshot(
+    `"2020-09-12T12:00:00.000Z UTC (MockDate: GMT+0000)"`
+  );
 });
