+++ packages/charts-utils/src/time/get-steps-from-interval/get-steps-from-interval.test.ts
@@ -1,0 +1,7 @@
+import getStepsFromInterval from './get-steps-from-interval';
+
+test('extracts step range from interval', () => {
+  const interval = 'every_20_days';
+
+  expect(getStepsFromInterval(interval)).toEqual(20);
+});
