+++ packages/charts/src/components/ruler/utils/generate-ruler/index.ts
@@ -1,0 +1,3 @@
+import generateRuler from './generate-ruler';
+
+export default generateRuler;
