+++ packages/charts/src/components/ruler/utils/index.ts
@@ -1,0 +1,4 @@
+import rotateLabel from './rotate-label';
+import generateRuler from './generate-ruler';
+
+export { rotateLabel, generateRuler };
