+++ packages/charts/src/components/ruler/utils/rotate-label/index.ts
@@ -1,0 +1,3 @@
+import rotateLabel from './rotate-label';
+
+export default rotateLabel;
