+++ packages/charts/src/components/ruler/utils/rotate-label/rotate-label.ts
@@ -1,65 +1,6 @@
-import { ScaleBand, ScaleLinear, ScaleTime } from 'd3-scale';
-import { ScaleSettings } from '@keen.io/charts-utils';
+import { Orientation, LabelRotation } from '../../../../types';
 
-import { generateTicks } from '../../utils/scale.utils';
-
-import { Orientation, LabelRotation } from '../../types';
-
-type Options = {
-  x: number;
-  y: number;
-  tickSize: number;
-  scale:
-    | ScaleBand<string>
-    | ScaleLinear<number, number>
-    | ScaleTime<number, number>;
-  orientation: Orientation;
-  scaleSettings?: ScaleSettings;
-  formatLabel?: (label: string | number) => string | number;
-};
-
-export const createRuler = ({
-  x,
-  y,
-  tickSize,
-  scale,
-  orientation,
-  scaleSettings,
-}: Options) => {
-  const [scaleStart, scaleEnd] = scale.range();
-  const ticks = generateTicks({
-    x,
-    y,
-    tickSize,
-    scale,
-    scaleSettings,
-    orientation,
-  });
-  let line;
-
-  if (orientation === Orientation.HORIZONTAL) {
-    line = {
-      x1: scaleStart,
-      x2: scaleEnd,
-      y1: y,
-      y2: y,
-    };
-  } else {
-    line = {
-      x1: x,
-      x2: x,
-      y1: scaleStart,
-      y2: scaleEnd,
-    };
-  }
-
-  return {
-    line,
-    ticks,
-  };
-};
-
-export const rotateLabel = (
+const rotateLabel = (
   orientation: string,
   radiusAngle: number,
   tickPadding: number,
@@ -150,6 +91,7 @@ export const rotateLabel = (
       }
     }
   }
+
   return {
     anchor: orientation === Orientation.HORIZONTAL ? 'middle' : 'end',
     radius: 0,
@@ -157,3 +99,5 @@ export const rotateLabel = (
     translateY: 0,
   };
 };
+
+export default rotateLabel;
